Add types module for Layer subtypes
